fix in config

+++ curltry/sObject.hpp
@@ -23,7 +23,7 @@ class sObject{
 private:
     std::string name;
     std::vector<sAttribute> attributeList {};
-    std::map<std::string,int> attributeCounters {};
+    std::map<std::string,long> attributeCounters {};
     void computeAttributes(const std::string &record, int);
 
 public:
